Add plugin
This is a basic calendar plugin I created for the InkyPi project. It features:
- Today’s Appointments: Displayed on the left side of the screen.
- ChatGPT Summary: A summary of the next three days’ appointments, generated using OpenAI’s API.

Feel free to modify or enhance it to fit your needs. Just keep in mind that this was a quick project, and some areas of the code might not be polished. Also, note that the screen content and some parts of the code are in German.

Requirements
- icalevents
- openai

In order to use this plugin, you need to install icalevents. Here's a instruction:

# 1. Stop InkyPi sevice
sudo systemctl stop inkypi.service
# 2. Switch to root so you can access the venv and run pip
sudo su
# 3. Activate the InkyPi virtual environment
source "/usr/local/inkypi/venv_inkypi/bin/activate"
# 4. Install the icalevents package in the venv
pip install icalevents
# 5. Deactivate the venv
deactivate
# 6. Exit root if you wish
exit
# 7. Finally, restart the InkyPi service
sudo systemctl restart inkypi.service

Important Notes:
You need a .ical URL (mandatory) and an OpenAI API Key (optional)
This plugin is made for the 7.3” Inky Impression 800x480 display. It may not work correctly on displays with other resolutions.
The plugin uses a stylish_calendar_config.json file to store settings. This file is automatically updated through the web interface, but you can also edit it manually if needed.
If you’re using ChatGPT summaries, your event details will be sent to OpenAI’s servers. Be mindful of the information you include in your calendar events.